refac: move components to folder, rename them and also improve some styles
