+++ src/app/components/dashboard/contributions-graph.tsx
@@ -3,9 +3,9 @@
 import CalendarHeatmap from 'react-calendar-heatmap';
 import 'react-calendar-heatmap/dist/styles.css';
 import { Tooltip } from 'react-tooltip';
-import { Contribution, Contributions } from "../api/contributions/route";
-import { shiftDateForward } from '../utils/index';
-import ContributionsLegend from './ContributionsLegend';
+import { Contribution, Contributions } from "../../api/contributions/route";
+import { shiftDateForward } from '../../utils/index';
+import ContributionsLegend from './contributions-legend';
 interface ContributionsProps {
     contributions: Contributions
 }
