+++ src/app/components/dashboard/contributions-legend.tsx
